backend/pyproject.toml
[project]
name = "creativeai-studio-backend"
version = "0.1.0"
description = "CreativeAI Studio - AI media generation platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-auth>=2.0.0",
    "sse-starlette>=2.0.0",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
    "greenlet>=3.3.2",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
